Reformat code to modularize tools
